Fix failure to build from source on GNU/kFreeBSD
GNU/Linux and GNU/kFreeBSD both use the GNU libc and runtime linker,
which is the layer that this engine interacts with directly. The specific
kernel used is below the level that this engine is dealing with.

From Debian, via ioquake3 commit
https://github.com/ioquake/ioq3/commit/d2f8b9f4bb3fda901ffd6f18074d035469832338.
Commit message added by Simon McVittie.

Bug-Debian: https://bugs.debian.org/413617
Reviewed-by: Simon McVittie <smcv@debian.org>
Signed-off-by: Simon McVittie <smcv@debian.org>