Short prox mines fuse feat applied to obelisks in Harvester
Apart from the different entity which identifies the bases (obelisk instead of flag), game mechanics of Harvester and OneFlag are similar in this regard (bring something to enemy base). So if the short mine exploding time (near your own base) is applied to OneFlag, it has got sense to also apply it to Harvester.