Added cg_zoomSensitivity for correct sens scaling
The zoom sensitivity scaling formula in OpenArena is `vertical_zoomFov / 75.0`.

This formula is flawed in 2 ways:

1. It assumes that the vertical unzoomed fov is 75. This is eqvivalent to
`cg_fov` being set to ~91 on 4:3 or ~107 on 16:9.

2. It scales the zoom sensitivity linearly using `vertical_zoomfov /
vertical_unzoomedfov`. This is mathematically incorrect which is noticable
through testing.

New cvar called `cg_zoomSensitivity` has been added to solve these problems.
When it is set to 0 (default) it has no effect and the legacy zoom itivity
formula is used.

When the cvar is set to anything but 0 the mathemtically correct formula
`tan(zoomFov / 2) / tan(unzoomedFov / 2)` is used. The result of this formula
gets multiplied by `cg_zoomSensitivity`.

When the cvar is set to 0.5 for example it will multiply the result of the
formula with 0.5 which halves the zoom sensitivity. This allows players to
customize their zoom sensitivity according to their preference.